refactor(desktop): move agentinspector parts to their owners
AgentInspector stopped being an inspector once the rail composition
(index.tsx) was deleted; it was a bag of parts under a stale name.
CostsSection moves beside AgentBrief, its only consumer. Resolver
parts (ChangesSection, BranchSurgery, thread cards, tests) move to
ResolverDetailPane/parts, matching the parts/ convention already
used by SessionNavSidebar and WorkflowBuilderView. AgentHeaderActions
is genuinely shared by AgentDetailPane and ResolverDetailPane, both
in features/session/components, so it becomes a flat file there
per file-system.md instead of moving into shared/components,
which is reserved for cross-feature code.
