+++ apps/desktop/src/features/session/components/AgentDetailPane/CostsSection.tsx
