+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/BranchSurgery.tsx
