+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/CommitRow.tsx
