+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverMissingVerdictsNotice.tsx
@@ -1,7 +1,7 @@
 import { GhostActionButton } from '@goodboy/ui';
-import { RESOLVER_ACTION_BUSY_LABEL } from '../../resolverActionBusyLabel';
-import { RESOLVER_ACTION_ICON } from '../../resolverActionIcon';
-import type { ResolverMissingVerdicts } from '../../resolverMissingVerdicts';
+import { RESOLVER_ACTION_BUSY_LABEL } from '../../../resolverActionBusyLabel';
+import { RESOLVER_ACTION_ICON } from '../../../resolverActionIcon';
+import type { ResolverMissingVerdicts } from '../../../resolverMissingVerdicts';
 
 type Props = {
   readonly missing: ResolverMissingVerdicts;
