+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverOverflowMenu.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { MoreHorizontal } from 'lucide-react';
 import { cn, Divider, Popover, useDropdown } from '@goodboy/ui';
 import type { Agent, BranchCommit } from '@goodboy/types';
-import { RESOLVER_ACTION_ICON } from '../../resolverActionIcon';
-import type { ResolverAction, ResolverActionKind } from '../../resolverActions';
-import type { ResolverActionsController } from '../../hooks/useResolverActions';
-import { ResolverConfirm } from '../ResolverConfirm';
+import { RESOLVER_ACTION_ICON } from '../../../resolverActionIcon';
+import type { ResolverAction, ResolverActionKind } from '../../../resolverActions';
+import type { ResolverActionsController } from '../../../hooks/useResolverActions';
+import { ResolverConfirm } from '../../ResolverConfirm';
 import { BranchSurgery } from './BranchSurgery';
 
 type Props = {
