+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverPanelSection.tsx
