+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverThreadCard.tsx
@@ -2,13 +2,13 @@ import { useState } from 'react';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import type { PrComment } from '@goodboy/types';
 import { GhostActionButton } from '@goodboy/ui';
-import type { ResolverActionKind } from '../../resolverActions';
-import type { ResolverRunningThreadAction } from '../../hooks/useResolverActions';
-import type { ResolverThreadSettlement } from '../../resolverThreadSettlements';
-import { resolverThreadBrief } from '../../resolverThreadBrief';
-import { resolverThreadDecisions } from '../../resolverThreadDecisions';
-import { resolverReplySummary } from '../../resolverReplySummary';
-import { prCommentLocation } from '../../pr-comment-location';
+import type { ResolverActionKind } from '../../../resolverActions';
+import type { ResolverRunningThreadAction } from '../../../hooks/useResolverActions';
+import type { ResolverThreadSettlement } from '../../../resolverThreadSettlements';
+import { resolverThreadBrief } from '../../../resolverThreadBrief';
+import { resolverThreadDecisions } from '../../../resolverThreadDecisions';
+import { resolverReplySummary } from '../../../resolverReplySummary';
+import { prCommentLocation } from '../../../pr-comment-location';
 import { ResolverThreadCardHeader } from './ResolverThreadCardHeader';
 import { ResolverThreadComment } from './ResolverThreadComment';
 import { ResolverThreadDecisions } from './ResolverThreadDecisions';
