+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverThreadComment.tsx
