+++ apps/desktop/src/features/session/components/ResolverDetailPane/parts/ResolverThreadReply.tsx
